+++ src/components/Toggle/index.tsx
@@ -3,15 +3,17 @@ import { Label, Slider } from './styles';
 
 export type ToggleProps = {
 	children?: ReactNode;
+  textFirst?: boolean;
 } & HTMLProps<HTMLInputElement>;
 
-const Checkbox = ({ children, className, style, ...props }: ToggleProps) => (
+const Checkbox = ({ children, textFirst, className, style, ...props }: ToggleProps) => (
 	<Label className={className} style={style}>
+    {textFirst && children}
 		<input type="checkbox" {...props} />
 		<Slider data-testid="slider">
 			<div />
 		</Slider>
-		{children}
+		{!textFirst && children}
 	</Label>
 );